WFC3 wf3ccd: Fix index variable type

+++ pkg/wfc3/lib/wf3ccd/dofwsat.c
@@ -69,7 +69,7 @@ int doFullWellSat(WF3Info *wf3, SingleGroup *x) {
     int DetCCDChip (char *, int, int, int *);
     void ComputeLimits(WF3Info *, int, int, int *, int *, int *, int *);
 
-    for (size_t i = 0; i < 2; i++) {
+    for (int i = 0; i < 2; i++) {
         xbeg[i] = -1;
         xend[i] = -1;
         ybeg[i] = -1;
@@ -222,10 +222,10 @@ int doFullWellSat(WF3Info *wf3, SingleGroup *x) {
     if (same_size) {
 
         /* Loop over the lines in the science image */
-        for (size_t j=ybeg[0]; j < yend[0]; j++) {
+        for (int j=ybeg[0]; j < yend[0]; j++) {
 
             /* Loop over the indices in the line in the science image */
-            for (size_t i = xbeg[0];  i < xend[0];  i++) {
+            for (int i = xbeg[0];  i < xend[0];  i++) {
                 /* Flag full-well saturated pixels with 256 dq bit*/             
                 if (Pix(x->sci.data, i, j) > (Pix(satimage.sci.data, i, j) / wf3->mean_gain)) {
                     sum_dq = DQPix(x->dq.data, i, j) | SATPIXEL;
@@ -234,7 +234,7 @@ int doFullWellSat(WF3Info *wf3, SingleGroup *x) {
             }
 
             /* If there is a second Amp in play, complete the processing of the line */ 
-            for (size_t i = xbeg[1];  i < xend[1];  i++) {
+            for (int i = xbeg[1];  i < xend[1];  i++) {
                 /* Flag full-well saturated pixels with 256 dq bit*/             
                 if (Pix(x->sci.data, i, j) > (Pix(satimage.sci.data, i, j) / wf3->mean_gain)) {
                     sum_dq = DQPix(x->dq.data, i, j) | SATPIXEL;
@@ -256,13 +256,13 @@ int doFullWellSat(WF3Info *wf3, SingleGroup *x) {
            l - line in reference image
         */
 
-        for (size_t j = 0, l = y0; j < ydim; j++, l++) {
+        for (int j = 0, l = y0; j < ydim; j++, l++) {
 
             /* Working with a subarray so need to apply the proper 
                section from the reference image to the science image.
             */
 
-            for (size_t i = 0, k = x0; i < xdim; i++, k++) {
+            for (int i = 0, k = x0; i < xdim; i++, k++) {
 
                 /* Increase the value of l to jump over the virtual overscan */
                 if (i == overstart)
